requirements.txt
fastapi
uvicorn
python-dotenv
supabase
sentence-transformers
pydantic
numpy
scipy
torch==2.2.0+cpu
-f https://download.pytorch.org/whl/cpu/torch_stable.html